fix: Fix clientInfo not being stored correctly
Before this commit, Client info stored with the Event Data is
being
looked up from the ClientInfo object which is stored in an
InheritableThreadLocal instance. Specifically its being looked up at the
time of processing the received event.
This InheritableThreadLocal
instance is populated in a httpfilter, and within the context of the
webflow is fine.
However, the Event Data is being posted via Spring
Async to a Spring eventListener.
Spring creates a ThreadExecuter to
handle these events,  and process them.
The executor is using a thread
pool that may not be created the thread within the Webflow.
This is
causing client data from other user's login requests being interwined
with the current user's principal, time of authentication.
This fix adds
the ClientInfo data into the 'event' objects being consume, and is now
being set at the time of being published, rather than at the time of
being received.  Since the publishEvent happens from the main webflow
thread, the data in the ClientInfo object should be valid.